backend/requirements.txt
# calmai backend api
# fastapi + async mongodb + auth + langchain rag

# web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# async mongodb
motor>=3.7.0

# auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# config
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.1.0

# langchain rag
langchain>=0.3.0
langchain-google-genai>=2.1.0
langchain-mongodb>=0.5.0
langchain-huggingface>=0.2.0

# embedding model
sentence-transformers>=3.0.0

# testing
httpx>=0.28.0
pytest>=8.0.0
pytest-asyncio>=0.25.0
pytest-cov>=6.0.0
